restructure quantum element module, enhance docstrings
