--- quafu/elements/quantum_element/pulses/__init__.py
@@ -1,11 +1,0 @@
-# !/usr/bin/env python
-# -*- coding:utf-8 -*-
-
-# @File: __init__.py.py
-# @Version: 1.0
-# @Author: Yun-Hao Shi
-# @Email: yhshi@iphy.ac.cn
-# @Reminders: |0> has been in the past, |1> is still in the future
-# @License: Copyright (c) 2023, IOP, CAS
-
-from .quantum_pulse import QuantumPulse
